refactor(api): move stage modification for linked target profile error to evaluation folder
