+++ api/lib/domain/errors.js
@@ -747,14 +747,6 @@ class TargetProfileInvalidError extends DomainError {
   }
 }
 
-class StageModificationForbiddenForLinkedTargetProfileError extends DomainError {
-  constructor(targetProfileId) {
-    super(
-      `Le profil cible ${targetProfileId} est déjà rattaché à une campagne. La modification du seuil ou niveau est alors impossible.`,
-    );
-  }
-}
-
 class OrganizationTagNotFound extends DomainError {
   constructor(message = 'Le tag de l’organization n’existe pas.') {
     super(message);
@@ -1167,7 +1159,6 @@ export {
   SessionNotAccessible,
   SessionWithIdAndInformationOnMassImportError,
   SIECLE_ERRORS,
-  StageModificationForbiddenForLinkedTargetProfileError,
   SupervisorAccessNotAuthorizedError,
   TargetProfileCannotBeCreated,
   TargetProfileInvalidError,
