+++ api/src/evaluation/application/http-error-mapper-configuration.js
@@ -1,6 +1,12 @@
 import { HttpErrors } from '../../shared/application/http-errors.js';
 import { DomainErrorMappingConfiguration } from '../../shared/application/models/domain-error-mapping-configuration.js';
-import { CompetenceResetError, EmptyAnswerError, ImproveCompetenceEvaluationForbiddenError, AcquiredBadgeForbiddenUpdateError } from '../domain/errors.js';
+import {
+  AcquiredBadgeForbiddenUpdateError,
+  CompetenceResetError,
+  EmptyAnswerError,
+  ImproveCompetenceEvaluationForbiddenError,
+  StageModificationForbiddenForLinkedTargetProfileError,
+} from '../domain/errors.js';
 
 const evaluationDomainErrorMappingConfiguration = [
   {
@@ -17,6 +23,12 @@ const evaluationDomainErrorMappingConfiguration = [
   },
   {
     name: CompetenceResetError.name,
+    httpErrorFn: (error) => {
+      return new HttpErrors.PreconditionFailedError(error.message);
+    },
+  },
+  {
+    name: StageModificationForbiddenForLinkedTargetProfileError.name,
     httpErrorFn: (error) => {
       return new HttpErrors.PreconditionFailedError(error.message);
     },
